feat(desktop): Desktop Support
Added Desktop Support
Imported dependencies
