+++ apps/desktopApp/src/main/kotlin/io/kmptemplate/desktop/main.kt
@@ -3,16 +3,13 @@ import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.material.Button
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 
+import io.kmptemplate.domain.Greeting
 @Composable
 @Preview
 fun App() {
@@ -20,7 +17,7 @@ fun App() {
 
     MaterialTheme {
         Button(
-            onClick = { text = "Hello, Desktop!" },
+            onClick = { text = Greeting().greet() },
             modifier = Modifier.testTag("button")
         ) {
             Text(text)
@@ -31,5 +28,6 @@ fun App() {
 fun main() = application {
     Window(onCloseRequest = ::exitApplication) {
         App()
+//        KmpTemplateApp()
     }
 }
