+++ shared/ui/base/src/commonMain/kotlin/io/kmptemplate/ui/base/design/ApplicationTheme.kt
