+++ shared/ui/main/src/commonMain/kotlin/io/kmptemplate/ui/main/view/GreetingView.kt
@@ -1,0 +1,9 @@
+package io.kmptemplate.ui.main.view
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun GreetingView(text: String) {
+    Text(text = text)
+}
